virtio_sound: fix: fixed queues, control messages, etc.
Control requests did not consider the reply buffer,
resulting in the queue always appearing to be full.

Fix several swapped I/O streams, where outputs and
inputs were mixed.

Fixed missing semaphores for the controlQueue.

If no chmaps are given for a stream, we now
default to a L/R channel mapping.

Signed-off-by: Diego Roux <diegoroux04@protonmail.com>